Move settings up a class level to DefaultTripalStringTypeFormatter
